# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Construction By Apex API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /opportunities/email-ingest:
    post:
      operationId: ingestEmail
      tags: [opportunities]
      summary: Ingest an inbound email as an opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailIngestBody"
      responses:
        "201":
          description: Email ingested
          content:
            application/json:
              schema:
                type: object
                properties:
                  inserted:
                    type: boolean
                  updated:
                    type: boolean
                required: [inserted, updated]
...
components:
  schemas:
...
    EmailIngestBody:
      type: object
      properties:
        from:
          type: string
        subject:
          type: string
        textBody:
          type: string
        htmlBody:
          type: string
        links:
          type: array
          items:
            type: string
        attachmentNames:
          type: array
          items:
            type: string
        receivedAt:
          type: string
          format: date-time