# lib/api-zod/src/generated/types/emailIngestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Construction By Apex API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailIngestBody {
  from?: string;
  subject?: string;
  textBody?: string;
  htmlBody?: string;
  links?: string[];
  attachmentNames?: string[];
  receivedAt?: Date;
}